chore: Reorganizes project layout
The project layout was very... flat. Most files for the project were just
slopped into `Assets`, without any subdivision. While this is a tiny demo
project, it bothered me that everything was just plopped into that one
space.

We now have materials _actually_ in the Materials folder! I also moved
textures to Textures, and moved all code to the Src folder.

The main scene is still in Assets. There are no plans for an additional
scene for this project, so there is really no need to make a separate
folder for it (can always be done later if we find the need).

One thing that is kind of hanging out is the "shittycamera" asset. If I
remember correctly, this was supposed to be a very basic model of a
camera to show where it is in the scene view. Upon loading this project
in a more modern version of unity, it seems this model is no longer
loading correctly. It could be the modern version, or it could be that
I made a mistake when adding it to the repo originally, which corrupted
it. This will be fixed in a followup commit.