build: Updated to target Unity 2022.1.13
The original target of this project was Unity 5, which at this point is pretty
old. I wanted to ensure that newer users who may stumble upon this demo are
definitely able to run it with less headaches. I expect the target audience of
this to be novices or beginners.

As it turns out, it wasn't a challenge whatsoever to load this project in this
version of Unity, but I'll save the trouble for people in the future and
supply this upgrade here.