fix: Resolves inability to build project in latest Unity version
I was getting the following error when trying to build the Project since
upgrading to the latest Unity version:

    Component GUI Layer in Main Camera for Scene Assets/main.unity is no
    longer available. It will be removed after you edit this GameObject
    and save the Scene.

As it turns out, versions of Unity since I first built this project no
longer come with certain components any longer. the root cause was a
"Flare Renderer" on the Main Camera. Simply removing this component was
all that was necessary to be able to build the project.

We never even needed it for this, so no harm no foul.