chore: Reorganized objects and removed unused ones
This sounds like more than it was. As it turns out, the "shittycamera" object
which was broken in a prior version of Unity seems to be working in the newer
version. This is a happy accident!

I also realized there was a dead asset of some sort in here, some old copy of
the shittycamera object that was unused. This got nuked.