docs: Adds a README worth reading
The original README.md in this project was, admittedly, pretty useless.
I am not sure how anyone would possibly know what this is with the
README.md I had previously.

The new README.md provides actual details about the project, with
pictures! Everyone loves pictures. More importantly, it gives a better
clue as to what this does, how to build it, and where to read more about
it.